Upgrade to Emscripten 3.1.27
This is the first release that's based on Ubuntu 22.04 LTS (jammy) that upgrades CMake from 3.16.3 to 3.22.1.

This is newer than emscripten 3.1.14 which is the version listed on https://doc.qt.io/qt-6/wasm.html

This also gives a new Emscripten toolchain location. Also, backports are no longer needed after the Ubuntu upgrade.